published items referencing preview only wip
