+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Contexts/GraphMergeContext.cs
@@ -1,27 +1,19 @@
 ﻿using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.Contexts;
-using DFC.ServiceTaxonomy.GraphSync.OrchardCore.Interfaces;
-using DFC.ServiceTaxonomy.GraphSync.OrchardCore.Wrappers;
 using DFC.ServiceTaxonomy.Neo4j.Commands.Interfaces;
 using DFC.ServiceTaxonomy.Neo4j.Services.Interfaces;
 using OrchardCore.ContentManagement;
-using OrchardCore.ContentManagement.Metadata.Models;
 
 namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Contexts
 {
     //todo: common base class for GraphMergeContext and ValidateAndRepairContext
     public class GraphMergeContext : GraphOperationContext, IGraphMergeItemSyncContext
     {
-        public IGraphSyncHelper GraphSyncHelper { get; }
         public IGraphReplicaSet GraphReplicaSet { get; }
         public IMergeNodeCommand MergeNodeCommand { get; }
         public IReplaceRelationshipsCommand ReplaceRelationshipsCommand { get; }
 
         public ContentItem ContentItem { get; }
-        public IContentManager ContentManager { get; }
-        public IContentItemVersion ContentItemVersion { get; }
-        public ContentTypePartDefinition ContentTypePartDefinition { get; set; }
-        public IContentPartFieldDefinition? ContentPartFieldDefinition { get; private set; }
 
         public IGraphMergeContext? ParentGraphMergeContext { get; }
 
@@ -34,24 +26,15 @@ namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Contexts
             IContentManager contentManager,
             IContentItemVersionFactory contentItemVersionFactory,
             IGraphMergeContext? parentGraphMergeContext)
+        : base(graphSyncHelper, contentManager)
         {
-            GraphSyncHelper = graphSyncHelper;
             GraphReplicaSet = graphReplicaSet;
             MergeNodeCommand = mergeNodeCommand;
             ReplaceRelationshipsCommand = replaceRelationshipsCommand;
             ContentItem = contentItem;
-            ContentManager = contentManager;
             ParentGraphMergeContext = parentGraphMergeContext;
 
             ContentItemVersion = contentItemVersionFactory.Get(graphReplicaSet.Name);
-
-            ContentTypePartDefinition = default!;
-        }
-
-        public void SetContentPartFieldDefinition(ContentPartFieldDefinition? contentPartFieldDefinition)
-        {
-            ContentPartFieldDefinition = contentPartFieldDefinition != null
-                ? new ContentPartFieldDefinitionWrapper(contentPartFieldDefinition) : default;
         }
     }
 }
