+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/Contexts/IGraphOperationContext.cs
@@ -1,7 +1,18 @@
-﻿namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.Contexts
+﻿using DFC.ServiceTaxonomy.GraphSync.OrchardCore.Interfaces;
+using OrchardCore.ContentManagement;
+using OrchardCore.ContentManagement.Metadata.Models;
+
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces.Contexts
 {
     public interface IGraphOperationContext
     {
+        IContentManager ContentManager { get; }
+        IContentItemVersion ContentItemVersion { get; }
+        ContentTypePartDefinition ContentTypePartDefinition { get; }
+        IContentPartFieldDefinition? ContentPartFieldDefinition  { get; }
 
+        IGraphSyncHelper GraphSyncHelper { get; }
+
+        void SetContentPartFieldDefinition(ContentPartFieldDefinition? contentPartFieldDefinition);
     }
 }
